--- src/tactmon/backend/db/orm/QueryGenerator.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-namespace backend::db::orm {
-    struct QueryGenerator {
-        QueryGenerator() { }
-
-        void Append(std::string_view value) {
-            _value += value;
-        }
-
-
-
-    private:
-        std::string _value;
-        std::unordered_map<uint32_t, std::string> _parameterNames;
-    };
-}
