+++ src/tactmon/backend/db/orm/select/CTE.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "backend/db/orm/concepts/Concepts.hpp"
+#include "backend/db/orm/Concepts.hpp"
 #include "utility/Literal.hpp"
 
 #include <ostream>
@@ -15,7 +15,7 @@ namespace backend::db::orm::select {
      * @tparam QUERY     The query if this CTE.
      */
     template <utility::Literal ALIAS, bool RECURSIVE, concepts::StreamRenderable QUERY>
-    struct CTE {
+    struct CTE final {
         template <size_t I>
         static auto render_to(std::ostream& ss, std::integral_constant<size_t, I>);
     };
